+++ modules/versions.lua
@@ -4,7 +4,7 @@
 return {
   ["Gen1Arena"] = "0.24.0",
   ["Gen1BattleUI"] = "1.8.0",
-  ["Gen1BillsBox"] = "1.8.1",
+  ["Gen1BillsBox"] = "1.8.2",
   ["Gen1Dex"] = "1.11.2",
   ["Gen1Elevator"] = "maintained",
   ["Gen1ItemInfo"] = "maintained",